Raised the warnings level in the introjucer, and fixed a few warnings that were thrown up by this.

+++ modules/juce_audio_basics/effects/juce_Reverb.h
@@ -333,7 +333,7 @@ private:
                 if (countdown <= 0)
                     currentValue = target;
                 else
-                    step = (target - currentValue) / countdown;
+                    step = (target - currentValue) / (float) countdown;
             }
         }
